moved websocket logic
