+++ experimenter_platform_stage_1_demo.py
@@ -1,5 +1,3 @@
-from StdSuites import null
-
 import tornado
 from tornado.options import define, options
 import os.path
@@ -36,14 +34,12 @@ class Application(tornado.web.Application):
             (r"/MMDIntervention", MMDInterventionHandler),
             (r"/questionnaire", QuestionnaireHandler),
             (r"/saveCoordinates", AjaxHandler),
-            (r"/writePolygon", PolygonAjaxHandler),
             (r"/resume", ResumeHandler),
             (r"/websocket", WebSocketHandler),
 
         ]
         #connects to database
         self.conn = sqlite3.connect('database.db')
-        self.conn2 = sqlite3.connect(params.USER_MODEL_STATE_PATH)
     #"global variable" to save current UserID of session
         UserID = -1;
         #global variable to track start and end times
@@ -330,22 +326,6 @@ class AjaxHandler(tornado.web.RequestHandler):
         file = open('application/frontend/static/AOICoordinates/'+jsonobj['filename'], 'w')
         file.write(self.request.body)
         file.close()
-
-
-class PolygonAjaxHandler(tornado.web.RequestHandler):
-    def post(self):
-        # gets polygon coordinates and refIds from frontend coordinateRefSentences
-        json_obj = json.loads(self.request.body)
-        for polygon_obj in json_obj:
-            ref_id = 'ref_' + polygon_obj.refId
-            polygon = polygon_obj.polygonCoords
-            polygon_data = (ref_id, self.application.cur_mmd, polygon, polygon)
-            # updates polygon in entry in db with same refId and task number
-            self.application.conn2.execute('INSERT INTO aoi(name, task, polygon) VALUES (?,?,?) ON DUPLICATE KEY UPDATE polygon=?', polygon_data)
-            # self.application.conn2.execute('UPDATE aoi SET polygon =? WHERE name =? AND task=?', polygon_data)
-        self.application.conn2.commit()
-        self.redirect('/')
-
 
 class PreStudyHandler(tornado.web.RequestHandler):
     def get(self):
